Refactor map and table controls into own component
