+++ src/components/LeafletMap.jsx
@@ -12,38 +12,11 @@ import '../customStyles.css'
 import customIcon from '../styles/Icons/map-pin.png'
 import usePrevious from '../library/usePrevious'
 import theme from '../theme'
-import styled, { css } from 'styled-components'
-import { ButtonSecondary } from './generic/buttons'
-import zoomToSelectedSites from '../styles/Icons/zoom_to_selected_sites.svg'
-import zoomToFiltered from '../styles/Icons/zoom_to_filtered.svg'
-import { mediaQueryTabletLandscapeOnly } from '../styles/mediaQueries'
+import MapAndTableControls from './MapAndTableControls'
 
 const defaultMapCenter = [32, -79]
 const defaultMapZoom = 2
 const mobileWidthThreshold = 960
-
-const ZoomToSecondaryButton = styled(ButtonSecondary)`
-  padding-left: 2rem;
-  padding-right: 2rem;
-  margin-left: 0.5rem;
-  display: flex;
-  flex-direction: row;
-  gap: 0.5rem;
-  align-items: center;
-`
-const ControlContainer = styled.div`
-  position: absolute;
-  top: 1.3rem;
-  left: 16.5rem;
-  height: 6rem;
-  z-index: 400;
-  display: flex;
-  flex-direction: row;
-  ${mediaQueryTabletLandscapeOnly(css`
-    top: 1rem;
-    left: 8rem;
-  `)}
-`
 
 const CircleMarkerPathOptions = {
   color: `${theme.color.white}`,
@@ -72,6 +45,9 @@ export default function LeafletMap(props) {
     setSelectedMarkerId,
     showFilterPane,
     showMetricsPane,
+    view,
+    setView,
+    projectDataCount,
   } = props
   const prevDisplayedProjects = usePrevious(displayedProjects)
   const location = useLocation()
@@ -99,55 +75,6 @@ export default function LeafletMap(props) {
     },
     [navigate, location.pathname],
   )
-
-  const handleZoomToFilteredData = () => {
-    if (!displayedProjects || displayedProjects.length === 0) {
-      return
-    }
-    const coordinates = displayedProjects.flatMap((project) =>
-      project.records.map((record) => [record.latitude, record.longitude]),
-    )
-    if (coordinates.length === 0) {
-      return
-    }
-    const bounds = L.latLngBounds(coordinates)
-    map.fitBounds(bounds)
-  }
-
-  const handleZoomToSelectedSite = () => {
-    const queryParams = new URLSearchParams(location.search)
-    if (queryParams.has('sample_event_id')) {
-      const sample_event_id = queryParams.get('sample_event_id')
-      const foundSampleEvent = displayedProjects
-        .flatMap((project) => project.records)
-        .find((record) => record.sample_event_id === sample_event_id)
-      if (!foundSampleEvent) {
-        return
-      }
-      const { latitude, longitude } = foundSampleEvent
-      map.setView([latitude, longitude], 18)
-    }
-  }
-
-  const isAnyActiveFilters = () => {
-    const queryParams = new URLSearchParams(location.search)
-    const filterKeys = [
-      'countries',
-      'organizations',
-      'startDate',
-      'endDate',
-      'method',
-      'dataSharing',
-      'projectName',
-    ]
-
-    return filterKeys.some((key) => queryParams.has(key))
-  }
-
-  const hasSelectedSite = () => {
-    const queryParams = new URLSearchParams(location.search)
-    return queryParams.has('sample_event_id')
-  }
 
   const toggleMapZoomControls = (map) => {
     if (window.innerWidth > mobileWidthThreshold) {
@@ -193,26 +120,13 @@ export default function LeafletMap(props) {
     toggleMapZoomControls(map)
 
     return (
-      <ControlContainer>
-        {isAnyActiveFilters() ? (
-          <ZoomToSecondaryButton
-            onClick={handleZoomToFilteredData}
-            aria-labelledby="Zoom to filtered data button"
-          >
-            <img src={zoomToFiltered} alt="Zoom to filtered data" />
-            {window.innerWidth > mobileWidthThreshold ? <span>Filtered Data</span> : null}
-          </ZoomToSecondaryButton>
-        ) : null}
-        {hasSelectedSite() ? (
-          <ZoomToSecondaryButton
-            onClick={handleZoomToSelectedSite}
-            aria-labelledby="Zoom to selected site button"
-          >
-            <img src={zoomToSelectedSites} alt="Zoom to selected site" />
-            {window.innerWidth > mobileWidthThreshold ? <span>Selected Site</span> : null}
-          </ZoomToSecondaryButton>
-        ) : null}
-      </ControlContainer>
+      <MapAndTableControls
+        map={map}
+        displayedProjects={displayedProjects}
+        projectDataCount={projectDataCount}
+        view={view}
+        setView={setView}
+      />
     )
   }
 
@@ -303,4 +217,7 @@ LeafletMap.propTypes = {
   setSelectedMarkerId: PropTypes.func.isRequired,
   showFilterPane: PropTypes.bool.isRequired,
   showMetricsPane: PropTypes.bool.isRequired,
+  view: PropTypes.oneOf(['mapView', 'tableView']).isRequired,
+  setView: PropTypes.func.isRequired,
+  projectDataCount: PropTypes.number.isRequired,
 }
